+++ test/audits/low-contrast-test.js
@@ -74,6 +74,27 @@ test("Disabled button = no relevant elements", function() {
     { result: axs.constants.AuditResult.NA });
 });
 
+test("aria-disabled button = no relevant elements", function() {
+  var fixture = document.getElementById('qunit-fixture');
+  var button = document.createElement('button');
+  button.textContent = "I Can Has Cheezburger?";
+  button.setAttribute("aria-disabled", "true");
+  fixture.appendChild(button);
+  deepEqual(axs.AuditRules.getRule('lowContrastElements').run({ scope: fixture }),
+    { result: axs.constants.AuditResult.NA });
+});
+
+test("aria-disabled=false button = pass", function() {
+  var fixture = document.getElementById('qunit-fixture');
+  var button = document.createElement('button');
+  button.textContent = "I Can Has Cheezburger?";
+  button.setAttribute("aria-disabled", "false");
+  fixture.appendChild(button);
+  deepEqual(
+    axs.AuditRules.getRule('lowContrastElements').run({ scope: fixture }),
+    { elements: [], result: axs.constants.AuditResult.PASS });
+});
+
 test("Button in disabled fieldset = no relevant elements", function() {
   var fixture = document.getElementById('qunit-fixture');
   var container = document.createElement('fieldset');
@@ -87,7 +108,41 @@ test("Button in disabled fieldset = no relevant elements", function() {
     { result: axs.constants.AuditResult.NA });
 });
 
+test("Button in disabled=false fieldset = no relevant elements", function() {
+  var fixture = document.getElementById('qunit-fixture');
+  var container = document.createElement('fieldset');
+  container.setAttribute("disabled", "false");  // check that the value of disabled is irrelevant
+  var legend = container.appendChild(document.createElement('legend'));
+  legend.textContent = 'Pointless legend';
+  var button = container.appendChild(document.createElement('button'));
+  button.textContent = "I Can Has Cheezburger?";
+  fixture.appendChild(container);
+  deepEqual(axs.AuditRules.getRule('lowContrastElements').run({ scope: fixture }),
+    { result: axs.constants.AuditResult.NA });
+});
+
+test("Button in aria-disabled container = no relevant elements", function() {
+  var fixture = document.getElementById('qunit-fixture');
+  var container = document.createElement('div');
+  container.setAttribute("role", "group");
+  container.setAttribute("aria-disabled", "true");
+  var legend = container.appendChild(document.createElement('legend'));
+  legend.textContent = 'Pointless legend';
+  var button = container.appendChild(document.createElement('button'));
+  button.textContent = "I Can Has Cheezburger?";
+  fixture.appendChild(container);
+  deepEqual(axs.AuditRules.getRule('lowContrastElements').run({ scope: fixture }),
+    { result: axs.constants.AuditResult.NA });
+});
+
 test("Crazy button in disabled fieldset legend test = relevant elements", function() {
+  /*
+   * Test this bit of the spec:
+   * The disabled attribute, when specified, causes all the form control descendants of the fieldset element,
+   * excluding those that are descendants of the fieldset element's first legend element child, if any,
+   * to be disabled.
+   * @see http://www.w3.org/TR/html5/forms.html#attr-fieldset-disabled
+   */
   var fixture = document.getElementById('qunit-fixture');
   var container = document.createElement('fieldset');
   container.setAttribute("disabled", "disabled");
@@ -102,6 +157,13 @@ test("Crazy button in disabled fieldset legend test = relevant elements", functi
 });
 
 test("Even crazier button in disabled fieldset second legend test = no relevant elements", function() {
+  /*
+   * Test this bit of the spec:
+   * The disabled attribute, when specified, causes all the form control descendants of the fieldset element,
+   * excluding those that are descendants of the fieldset element's first legend element child, if any,
+   * to be disabled.
+   * @see http://www.w3.org/TR/html5/forms.html#attr-fieldset-disabled
+   */
   var fixture = document.getElementById('qunit-fixture');
   var container = document.createElement('fieldset');
   container.setAttribute("disabled", "disabled");
@@ -114,4 +176,21 @@ test("Even crazier button in disabled fieldset second legend test = no relevant 
   fixture.appendChild(container);
   deepEqual(axs.AuditRules.getRule('lowContrastElements').run({ scope: fixture }),
     { result: axs.constants.AuditResult.NA });
+});
+
+test("Low contrast, disabled on undisableable = fail", function() {
+  var fixture = document.getElementById('qunit-fixture');
+  var div = document.createElement('div');
+  div.setAttribute("disabled", "disabled");
+  div.setAttribute("role", "checkbox");
+  div.tabIndex = 0;
+  div.disabled = true;
+  div.style.backgroundColor = 'white';
+  div.style.color = 'white';
+  div.textContent = 'Some text';
+  fixture.appendChild(div);
+  deepEqual(
+    axs.AuditRules.getRule('lowContrastElements').run({ scope: fixture }),
+    { elements: [div], result: axs.constants.AuditResult.FAIL }
+  );
 });